backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
PyJWT==2.6.0
cryptography==40.0.2
pyotp==2.8.0
django-environ==0.10.0
bcrypt==4.0.1
